Разделяй и влавствуй
